services/evaluation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers>=2.3.0
huggingface_hub>=0.20.0
httpx==0.25.0
pydantic==2.5.0
python-dateutil==2.8.2
numpy==1.24.3
